Moved BrowserLauncher to govkit - wip
