+++ gradle/verification-metadata.xml
@@ -979,6 +979,9 @@
          </artifact>
       </component>
       <component group="androidx.compose.ui" name="ui" version="1.0.1">
+         <artifact name="ui-1.0.1.aar">
+            <sha256 value="1943daa4a3412861b9a2bdc1a7c8c2ff05d9b8191c1d3e56ebb223d2eb4a8526" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
          <artifact name="ui-1.0.1.module">
             <sha256 value="57031a6ac9b60e5b56792ebf5cde6e16812ff566ed9190cbd188b00b46c13779" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
@@ -1020,6 +1023,14 @@
             <sha256 value="d972b219d5a08046c9e6af8a6a616b86e1c87a54aaa9effa78552f05729f456c" origin="Generated by Gradle"/>
          </artifact>
       </component>
+      <component group="androidx.compose.ui" name="ui-geometry" version="1.0.1">
+         <artifact name="ui-geometry-1.0.1.aar">
+            <sha256 value="04dcea79de0af8347f4517bec20578d4240ae5ee40191020ca07d7a38e640033" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+         <artifact name="ui-geometry-1.0.1.module">
+            <sha256 value="e2fea4b568ff93e340bc4b18c8a3c2a9bafdf3e1210d9ed30036c1d9888be369" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+      </component>
       <component group="androidx.compose.ui" name="ui-geometry" version="1.11.2">
          <artifact name="ui-geometry-1.11.2.module">
             <sha256 value="a87bbbbe13af56b9c0c514d25ffb8bccdc419f58eb3ac745b44b2f0835cf3b6a" origin="Generated by Gradle"/>
@@ -1050,6 +1061,14 @@
          </artifact>
          <artifact name="ui-geometry.aar">
             <sha256 value="baa21c336bfc9a7daf5ad32d50c5d436c92a8ac3492210d53900a1b395af4e3d" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
+      <component group="androidx.compose.ui" name="ui-graphics" version="1.0.1">
+         <artifact name="ui-graphics-1.0.1.aar">
+            <sha256 value="6b7f5d0c944483ea390b3d6da8438e7a69e0cfe310bd16c3c5e78ac0015b1663" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+         <artifact name="ui-graphics-1.0.1.module">
+            <sha256 value="ad9ce40deec721b8988c43ab847d803d00bee88c67cfc838dee565691f35db95" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="androidx.compose.ui" name="ui-graphics" version="1.11.2">
@@ -1156,6 +1175,14 @@
          </artifact>
          <artifact name="ui-test-manifest-1.9.5.module">
             <sha256 value="0a900388ca8a084e5a8a9c8e17a2ab814dbc7982e7862c25e80eb4cd22256122" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
+      <component group="androidx.compose.ui" name="ui-text" version="1.0.1">
+         <artifact name="ui-text-1.0.1.aar">
+            <sha256 value="f276b403fb95bda29fc83a33ff717591662ea2c769fc34c2077ec460b65f0da0" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+         <artifact name="ui-text-1.0.1.module">
+            <sha256 value="43a1ec94d3cecc5f3bd20f1c53551ab03a0ba728af1871a74bf239c418795f1b" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="androidx.compose.ui" name="ui-text" version="1.11.2">
@@ -1286,6 +1313,14 @@
             <sha256 value="64e6e6fb831b701241e9e19a743686f49c897ac806f122341d3051b6b82c84e8" origin="Generated by Gradle"/>
          </artifact>
       </component>
+      <component group="androidx.compose.ui" name="ui-unit" version="1.0.1">
+         <artifact name="ui-unit-1.0.1.aar">
+            <sha256 value="75b13e2a15da8a31acf75d81c4277ff2820b1bf58291b4688ac55200cdc18072" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+         <artifact name="ui-unit-1.0.1.module">
+            <sha256 value="3db12e5b47d6f6da91ed46feb902fe511282c846e4ef1affbb914779fa6b1082" origin="Generated by Gradle" reason="Artifact is not signed"/>
+         </artifact>
+      </component>
       <component group="androidx.compose.ui" name="ui-unit" version="1.11.2">
          <artifact name="ui-unit-1.11.2.module">
             <sha256 value="9dc285e5cce4f7dd8f1d1ea6d2b2d3e29e8dcdd157989fb43ca08f8ce97d08ac" origin="Generated by Gradle"/>
@@ -1316,6 +1351,11 @@
          </artifact>
          <artifact name="ui-unit.aar">
             <sha256 value="389a3c21d1ca0f89679d09e054db3aac72aa10630b2d2d2d4b46a81340020a23" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
+      <component group="androidx.compose.ui" name="ui-util" version="1.0.1">
+         <artifact name="ui-util-1.0.1.module">
+            <sha256 value="6fdae8e4b8785034255fc907b0fff5e61eed0672de4fff15b608614e3a8b086f" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="androidx.compose.ui" name="ui-util" version="1.11.2">
@@ -1907,6 +1947,11 @@
          </artifact>
          <artifact name="lifecycle-common-java8-2.10.0.module">
             <sha256 value="45328f75ac077a36381dc53d68c05288ba198137cabc41fdd4d24dc447eeddba" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
+      <component group="androidx.lifecycle" name="lifecycle-common-java8" version="2.3.0">
+         <artifact name="lifecycle-common-java8-2.3.0.module">
+            <sha256 value="2de56435ecd68ba2d49ae48011c34b758304af0a8e7e09ad2e8edfd2fe349b9c" origin="Generated by Gradle" reason="Artifact is not signed"/>
          </artifact>
       </component>
       <component group="androidx.lifecycle" name="lifecycle-common-java8" version="2.6.1">
